Very Large Update, See Notes
 - This commit completely breaks the server architecture of the client software. I'm
    working on fixing this.
 - Replace AlexaComp._log.XXX() calls with clog() calls.
 - Moved AudioController.cs to Controllers dir
 - Small updates to the compatibility tool
 - Moved forms to Forms dir
 - Moved TestFunctions.cs to Tests dir
 - Create AlexaCompCore class
    - All AlexaComp Classes except a few are now derived from AlexaCompCore.
    - Holds methods and values used across most or all AlexaComp classes.
 - AlexaComp.cs
    - Still serves as application entry point.
    - Moved many methods and variables to AlexaCompCore
 - AlexaCompGUI.cs
    - Shuffled methods around. File structure now makes more sense.
    - Added region tags.
 - AlexaCompHARDWARE.cs
    - Added getSensorType method.
    - Modified some of the loops a bit.
 - AlexaCompREQUEST.cs
    - Modified Response class to take default params.
    - Modified Response constructions to reflect addition of default params.
    - Refactored audioRequest()
    - Refactored processRequest()
 - AlexaCompSERVER.cs
    - Added AlexaCompServerInstance class.
        - AlexaCompServerInstance will now handle receiving, deserializing, and
          processing requests. New server is created on class construction.
