+++ AlexaComp/AlexaCompREQUEST.cs
@@ -2,10 +2,11 @@
 using System.Xml;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
+using AlexaComp.Controllers;
 using System.Threading;
 
 namespace AlexaComp {
-    public class AlexaCompREQUEST {
+    class AlexaCompREQUEST : AlexaCompCore{
 
         [DllImport("user32.dll")]
         private static extern bool ExitWindowsEx(uint uFlags, uint dwReason);
@@ -16,116 +17,89 @@ namespace AlexaComp {
         private static string failMessage = "There was an error, please check the Alexa Comp log file.";
         public static void processRequest(Request req) {
             switch (req.COMMAND) {
-                case "LAUNCH":
-                    launchRequest(req);
-                    req.logTimeElapsed();
-                    break;
-                case "COMMAND":
-                    commandRequest(req);
-                    req.logTimeElapsed();
-                    break;
-                case "GETCOMPSTAT":
-                    compStatRequest(req);
-                    req.logTimeElapsed();
-                    break;
-                case "AUDIOCOMMAND":
-                    audioRequest(req);
-                    break;
-                case "NETWORKCOMMAND":
-                    break;
+                case "LAUNCH":         launchRequest(req); break;
+                case "COMMAND":        commandRequest(req); break;
+                case "GETCOMPSTAT":    compStatRequest(req); break;
+                case "AUDIOCOMMAND":   audioRequest(req); break;
+                case "NETWORKCOMMAND": break; // TODO : Implement NWCommand
+                case "RGBCOMMAND":     break; // TODO : Implement RGBCommand
             }
+            req.logTimeElapsed();
         }
+
+        /*
+        static void RGBRequest(Request req) {
+            switch (req.PRIMARY) {
+                case "RAINBOWFADE": LightingController.LightingEffects.fadingEffect(); break;
+                case "STATICCOLOR": LightingController.LightingEffects.staticColor(255, 255, 255); break;
+            }
+        }*/
 
         static void audioRequest(Request req) {
             try {
                 switch (req.PRIMARY) {
-                    case "PLAYPAUSE":
-                        AudioController.togglePlayPause();
-                        break;
-                    case "NEXTTRACK":
-                        AudioController.nextTrack();
-                        break;
-                    case "PREVTRACK":
-                        AudioController.prevTrack();
-                        break;
-                    case "SETVOLUME":
-                        AudioController.setVolume(Convert.ToInt32(req.SECONDARY));
-                        break;
-                    case "VOLUMEUP":
-                        AudioController.volUp();
-                        break;
-                    case "VOLUMEDOWN":
-                        AudioController.volDown();
-                        break;
-                    case "TOGGLEMUTE":
-                        break;
+                    case "PLAYPAUSE":  AudioController.togglePlayPause(); break;
+                    case "NEXTTRACK":  AudioController.nextTrack(); break;
+                    case "PREVTRACK":  AudioController.prevTrack(); break;
+                    case "SETVOLUME":  AudioController.setVolume(Convert.ToInt32(req.SECONDARY)); break;
+                    case "VOLUMEUP":   AudioController.volUp(); break;
+                    case "VOLUMEDOWN": AudioController.volDown(); break;
+                    case "TOGGLEMUTE": break; // TODO: Implement Mute Functionality
                 }
-                Response res = new Response(true, "Done!", "", "");
+                Response res = new Response(true, "Done!");
             } catch (Exception e) {
-                AlexaComp._log.Error(e);
-                Response res = new Response(false, "Oops! Something went wrong...", "", "");
+                clog(e.ToString());
+                Response res = new Response(false, "Oops! Something went wrong...");
             }
         }
 
         static void launchRequest(Request req) {
             if (req.PRIMARY == "SHUTDOWN") {
-                AlexaCompSERVER.stopServer();
+                // AlexaCompSERVER.stopServer();
                 Process.Start("shutdown", "/s /t .5");
             } 
 
             // For starting a
             try {
                 string programPath = GetProgramPath(req.PRIMARY);
-                AlexaComp._log.Info("ProgramPath - " + programPath);
-                AlexaComp._log.Info("req Program - " + req.PRIMARY);
+                clog("ProgramPath - " + programPath);
+                clog("req Program - " + req.PRIMARY);
                 Process.Start(GetProgramPath(req.PRIMARY));
-                AlexaComp._log.Info("Program Launched");
+                clog("Program Launched");
                 Response res = new Response(true, "Program Launched!", "", "");
-                AlexaCompSERVER.stopServer(); // Restart Server to Handle Next Request
+                // AlexaCompSERVER.stopServer(); // Restart Server to Handle Next Request
             }
             catch (NullReferenceException) {
-                AlexaComp._log.Error("NullReferenceException caught during attempt to launch program, " +
+                clog("NullReferenceException caught during attempt to launch program, " +
                     "null most likely returned from GetProgramPath.");
-                Response res = new Response(false, failMessage, "", "");
-                AlexaCompSERVER.stopServer();
+                Response res = new Response(false, failMessage);
+                // AlexaCompSERVER.stopServer();
             }
             catch (System.ComponentModel.Win32Exception e) {
-                AlexaComp._log.Error("Win32Exception Caught during attempt to launch program " + e.Message);
-                Response res = new Response(false, failMessage, "", "");
-                AlexaCompSERVER.stopServer();
+                clog("Win32Exception Caught during attempt to launch program " + e.Message);
+                Response res = new Response(false, failMessage);
+                // AlexaCompSERVER.stopServer();
             }catch (InvalidOperationException e) {
-                AlexaComp._log.Error("InvalidOperationException Caught during attempt to launch program " + e.Message);
-                Response res = new Response(false, failMessage, "", "");
-                AlexaCompSERVER.stopServer();
+                clog("InvalidOperationException Caught during attempt to launch program " + e.Message);
+                Response res = new Response(false, failMessage);
+                // AlexaCompSERVER.stopServer();
             }
         }
 
         static void commandRequest(Request req) {
             switch (req.PRIMARY) {
-                case "SHUTDOWN":
-                    AlexaComp.stopApplication();
-                    Process.Start("shutdown", "/s /t .5");
-                    break;
-                case "LOCK":
-                    Process.Start(@"C:\WINDOWS\system32\rundll32.exe", "user32.dll,LockWorkStation");
-                    break;
-                case "RESTART":
-                    AlexaComp.stopApplication();
-                    Process.Start("shutdown", "/r /t .5");
-                    break;
-                case "SLEEP":
-                    SetSuspendState(false, true, true);
-                    break;
-                case "LOGOFF":
-                    ExitWindowsEx(0, 0);
-                    break;
-                case "HIBERNATE":
-                    SetSuspendState(true, true, true);
-                    break;
+                case "SHUTDOWN":  stopApplication(); Process.Start("shutdown", "/s /t .5"); break;
+                case "LOCK":      Process.Start(@"C:\WINDOWS\system32\rundll32.exe", "user32.dll,LockWorkStation"); break;
+                case "RESTART":   stopApplication(); Process.Start("shutdown", "/r /t .5"); break;
+                case "SLEEP":     SetSuspendState(false, true, true); break;
+                case "LOGOFF":    ExitWindowsEx(0, 0); break;
+                case "HIBERNATE": SetSuspendState(true, true, true); break;
             }
         }
 
         static void compStatRequest(Request req) {
+            // TODO : Fix Compstat System
+            // TODO : Implement switch case
             if (req.PRIMARY == "RAM") {
                 Response res = new Response(true, "5.3 Gigabites", "", "");
             }
@@ -140,46 +114,50 @@ namespace AlexaComp {
         static string GetProgramPath(string program) {
             XmlDocument doc = new XmlDocument();
             doc.Load("pathDir.xml");
-            AlexaComp._log.Info("pathDIR loaded");
+            clog("pathDIR loaded");
             XmlElement elem = doc.SelectSingleNode("//path[@programName='" + program + "']") as XmlElement;
             if (elem != null) {
                 string path = elem.GetAttribute("programPath");
-                AlexaComp._log.Info("Path - " + path);
+                clog("Path - " + path);
                 return path;
             }
             else {
-                AlexaComp._log.Info("null returned");
+                clog("null returned");
                 return null;
             }
         }
     }
 
-    public class Request {
+    [DebuggerDisplay("[AlexaComp Request Object -- {{Command: {COMMAND}, Primary: {PRIMARY}, Secondary: {SECONDARY}, Tertiary: {TERTIARY}}}]")]
+    class Request : AlexaCompCore{
         public string AUTH { get; set; }
         public string COMMAND { get; set; }
         public string PRIMARY { get; set; }
         public string SECONDARY { get; set; }
         public string TERTIARY { get; set; }
 
+        public static AlexaCompServerInstance originInstance;
+
         public Stopwatch sw = new Stopwatch();
 
-        public Request() {
+        public Request(AlexaCompServerInstance inst) {
             sw.Start();
-            printRequest();
+            originInstance = inst;
         }
 
         public void logTimeElapsed() {
             sw.Stop();
-            AlexaComp._log.Info("Request Completed, Time Elapsed(ms) - " + sw.ElapsedMilliseconds.ToString());
+            clog("Request Completed, Time Elapsed(ms) - " + sw.ElapsedMilliseconds.ToString());
         }
 
+        /*
         public void printRequest() {
             string str = string.Format("New Request - {{Command: {0}, Primary: {1}, Secondary: {2}, Tertiary: {3}}}", COMMAND, PRIMARY, SECONDARY, TERTIARY);
-            Console.WriteLine(str);
-        }
+            clog(str);
+        }*/
     }
 
-    public class Response {
+    class Response : AlexaCompCore{
         public bool passorfail;
         public string message;
         public string primary;
@@ -187,7 +165,7 @@ namespace AlexaComp {
 
         public string json;
 
-        public Response(bool passorfail_, string message_, string primary_, string secondary_) {
+        public Response(bool passorfail_, string message_, string primary_ = "", string secondary_ = "") {
             passorfail = passorfail_;
             message = message_;
             primary = primary_;
@@ -199,9 +177,9 @@ namespace AlexaComp {
 
         public void sendResponse(System.Net.Sockets.NetworkStream customStream) {
             if (customStream != null) {
-                AlexaCompSERVER.sendToLambda(json, customStream);
+                // AlexaCompSERVER.sendToLambda(json, customStream);
             } else {
-                AlexaCompSERVER.sendToLambda(json, null);
+                // AlexaCompSERVER.sendToLambda(json, null);
             }
         }
     }
