+++ AlexaComp/CompatibilityTool/bin/Debug/CompatibilityTool.exe.config
@@ -3,6 +3,7 @@
   <configSections>
     <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net" />
   </configSections>
+  
   <log4net>
     <appender name="LogFileAppender" type="log4net.Appender.RollingFileAppender">
       <param name="File" value="CompatibilityToolLOG.txt"/>
