+++ AlexaComp/packages.config
@@ -32,8 +32,10 @@
   <package id="RGB.NET.Resources.CoolerMaster" version="0.1.0" targetFramework="net461" />
   <package id="RGB.NET.Resources.Corsair" version="0.1.0" targetFramework="net461" />
   <package id="RGB.NET.Resources.Logitech" version="0.1.0" targetFramework="net461" />
+  <package id="RGB.NET.Resources.Msi" version="0.1.0" targetFramework="net461" />
   <package id="RGB.NET.Resources.Novation" version="0.1.0" targetFramework="net461" />
   <package id="RGB.NET.Resources.Razer" version="0.1.0" targetFramework="net461" />
   <package id="Sanford.Multimedia.Midi" version="6.4.1" targetFramework="net461" />
+  <package id="System.Management" version="4.5.0" targetFramework="net461" />
   <package id="System.ValueTuple" version="4.4.0" targetFramework="net461" />
 </packages>
